requirements.txt
pandas
requests
beautifulsoup4
selenium
webdriver-manager
streamlit
plotly
google-generativeai
schedule
python-dotenv